Refactored code to simply and test new structures to plate, well, sample and sampleframe.
Removed statistical and replicate redundancy across classes and moved them all to the sampleframe class.

The well class is now a simple container for well metadata and time series data.

The plate class is now a data container for the well class meant to organize and provide structured analysis for collections of well data.

The sample class is now a container class for replicate analysis of sample types across all concentrations.

The sampleframe is a container for sample objects.